Way too many things

+++ src/util/math/Vector.java
@@ -9,7 +9,7 @@ public abstract class Vector implements Serializable, ReadableVector {
     }
 
     public final float length() {
-        return (float) Math.sqrt((double) this.lengthSquared());
+        return (float) Math.sqrt(this.lengthSquared());
     }
 
     public abstract float lengthSquared();
